Move framed header to page headers
