+++ common/views/components/PageHeaders/PageHeaders.config.json
@@ -1,0 +1,3 @@
+{
+  "name": "Page headers"
+}
